Add tests for fine-grained notifications on linking objects properties.
The tests revealed that collection changes were not being correctly
computed for backlinks as `ResultsNotifier` was unaware of the ordering
of objects in linking objects properties. To address this `Results` can
now report whether it contains rows that are guaranteed to be in table
order. `ResultsNotifier` uses this information to determine when
`CollectionChangeBuilder` can take its optimized path rather than only
disabling it for `Results` built from `LinkView`s.